Rename files.
